backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
boto3==1.34.0
python-dotenv==1.0.0
Werkzeug==3.0.1
groq==0.4.1
tavily-python==0.3.0
google-generativeai==0.3.1

